feat: in-protocol minimum gas price (#107)
Resolves https://github.com/babylonlabs-io/pm/issues/49

This PR introduces the in-protocol minimum gas price mechanism, in which
the consensus (more specifically the AnteHandler) enforces every tx has
to set a gas price at least 0.002 ubbn. The PR also provides relevant
tests. The impl closely follows
https://github.com/celestiaorg/celestia-app/pull/2985.

In addition, this PR creates a new package `app/ante` to abstract out
the construction of the AnteHandler for Babylon, following the practice
at Celestia.

- RFC: https://github.com/babylonlabs-io/pm/pull/56
- ADR: https://github.com/babylonlabs-io/pm/pull/61