chore: Do not convert int to uint32 in btcstaking lib
This PR handles the conversion of ints to uint32s in the btcstaking
library. The conversions that were modified were not unsafe
as they were the results of the `len()` call,
but they still made static analyzers complain.

Instead, we opted to convert uint32s to ints. However,
this makes the assumption that the program will be executed in
64-bit architectures in order to be safe, which afaik is already
a requirement.

Further, the PR includes some other small conversion fixes
that have been found in the codebase -- mostly low hanging fruit.