chore: Remove nolint:errcheck and instead handle errors (#134)
As we are adding static analyzers in the codebase, many of them will
point out unhandled errors. If we add annotations to our codebase for
each of them, this might bloat the codebase. Another approach is to only
allow a single static analyzer fail about unhandled errors, while
ignoring the rest.

In this PR, I opted to just handle the errors by panicking in order to remove
the `nolint:errcheck` from the annotations. These methods failing
are programmatic errors.