chore: Remove localnet functionality from Makefile
This functionality is not useful as a network of Babylon nodes just by itself does not allow us to test much.
This is evident by the fact that we never use this for testing.
Instead we typically utilise a deployment containing all the interconnected components
consisting of the Babylon network.