chore: Bitcoin heights and depths made uint32 (#181)
Resolves #179 

This is an impactful change that brings the BTC height types that we use
more in sync with the expected types for BTC block heights, similar to
those that are used by other libraries. It is part of the code
improvement efforts that we are doing in preparation of the next
testnet.

Given that it touches many files, I believe that this change should
happen sooner than later so that it can be appropriately tested.