[backport] feat: BLS keystore improvement (#467) (#478)
This PR is backport from #467.

## Summary

### Description

Previously, the `priv_validator_key.json` file, created locally during
the chain initialization process, stored both the BLS keys and the
FilePV of Comet.

Due to this mixed structure, ExtendVote became dependent on the local
`priv_validator_key.json` file. As a result, it was unusable in
environments where the `priv_validator_key.json` file was not stored
locally, such as with a remote signer.

To address this issue, we introduced the migrate-bls-key command. This
command extracts the BLS key from `priv_validator_key.json` and stores
it in a separate file, ensuring that `priv_validator_key.json` retains
only the FilePV content generated during Comet’s initialization process.

Since the BLS key is stored in a file formatted according to ERC-2335, a
password is required during the migration process. This password is
stored in a separate file.

### Requirements

First, we assume that the node is operating normally.

If you run a node for the first time with a new version, the `babylond
init` command will create a `priv_validator_key.json` file that stores
Comet’s FilePV, along with a separate file for the BLS key.

Additionally, the node must be running a previous version that does not
use a separate file structure. If you are already using a newer version
that creates a separate file structure, running the `migrate-bls-key`
command will result in the following error:

```sh
failed to load previous version of priv_validator_key.json in ...
```

### Test migration-bls-key

#### 1. Setup and start the node using the `babylond` version, e.g.,
`v1.0.0-rc.4`.

```sh
$ babylond init my-node --chain-id test-chain-1
$ babylond keys add validator1 --keyring-backend test
$ babylond add-genesis-account $(babylond keys show validator1 -a --keyring-backend test) 2000000000000ubbn,2000000000000stake
$ babylond gentx validator1 1500000000000stake --chain-id test-chain-1 --keyring-backend test --fees 400ubbn
$ babylond collect-gentxs
$ babylond create-bls-key $(babylond keys show validator1 -a --keyring-backend test)
$ babylond gen-helpers create-bls
# copy content of gen-bls-x.json into checkpointing.genesis_keys field in genesis.json
$ babylond start
```

#### 2. Stop the running node.
#### 3. Execute migration from the changed version of `babylond`.

Before calling migration, check if the `priv_validator_key.json` file
storing the two types of keys exists inside `<home>/.babylond/config`.

```sh
$ babylond migrate-bls-key
```

#### 4. Enter the password you will use to encrypt the bls key into the
erc2335 structure.
#### 5. Verify that the `bls_key.json` and `bls_password.txt` files are
correctly created in `<home>/.babylond/config`.
#### 6. Verify that `priv_validator_key.json` contains only keys for
comet.
#### 7. Restart the node using changed version of `babylond`.

```sh
$ babylond start
```

#### Error cases
- If `priv_validator_key.json` does not exist when calling migration
- If `priv_validator_key.json` does not satisfy the previous version of
a structure containing comet and bls keys.

Co-authored-by: wonjoon <39115630+wnjoon@users.noreply.github.com>
Co-authored-by: dongsam <dongsamb@gmail.com>