chore: Remove async-icq module (#993)
This PR removes the wiring of the async ICQ module. It includes the
following:
* Remove wiring of the async-icq module on the app/app.go
* Introduce a v2rc1 upgrade handler for the testnet that removes the
module.
* Update the v2 upgrade handler to have the option to specify whether to
include the async-icq module or not. For the testnet upgrade, this has
been set to `true`, i.e., include the module, while for the mainnet
upgrade, this option has been set to `false.
* The option was added as the v2 upgrade has already been executed on
the testnet with the module included. In order to have a consistent
sequence of the upgrades for both testnet and mainnet, we need to
maintain the upgrade with the module included.
* An alternative option is to modify the upgrade to v2rc0 for the
testnet and copy its contents in a new v2 upgrade for the mainnet. This
seems to be repeating code and complicating things, so it was not
followed.
* Small updates to resolve a few inconsistencies (e.g., for
`x/tokenfactory` we use `ModuleName` instead of `StoreKey` in the
upgrade).

(cherry picked from commit a34969aaf73e0e051aa445149dfb4bb06255b10d)

# Conflicts:
#	CHANGELOG.md
#	app/keepers/keepers.go