feat: EVM inclusion (#1229)
This PR includes the minimum necessary changes to make the Babylon
Genesis chain EVM compatible. This includes:

- `x/vm` module
- `x/erc20` module
- `x/feemarket` module
-  Static precompiles for bank, gov, slashing, evidence, bech32 and p256
- Token pair only for the native BABY token 
- Antehandler for EVM transactions
- `x/precisebank` module for gas conversions
-  Included config for EVM and JSON-RPC for `testnet` command


## How to test

Start the chain with the `testnet` command:

```
 make build && ./build/babylond testnet \
    --v                     1 \
    --output-dir            ./.testnet \
    --starting-ip-address   192.168.10.2 \
    --keyring-backend       test \
    --chain-id              chain-test \
    --additional-sender-account true \
&& ./build/babylond start --home ./.testnet/node0/babylond --log_level=info
```

This will create a chain with a new `eth_secp` account that you can
import in Metamask with the following known dev mnemonic:

```
doll midnight silk carpet brush boring pluck office gown inquiry duck chief aim exit gain never tennis crime fragile ship cloud surface exotic patch
```

The configuration for adding a new chain to metamask is the following:

- **Network Name**: Babylon Genesis
- **RPC URL**: http://127.0.0.1:8545
- **Chain ID**: 6901 (can be modified to whatever we want here)
- **Currency Symbol**: BABY

## Known Issues

Due to limitations with the Cosmos EVM and the gas mechanism the EVM
does not currently work with our `PriorityNonceMempool` so the `NoOp`
mempool is configured in the `testnet` command. The issue was raised
with the Cosmos EVM team.

---------

Co-authored-by: RafilxTenfen <rafaeltenfen.rt@gmail.com>
Co-authored-by: Tom <54514587+GAtom22@users.noreply.github.com>