fix(zoneconcierge): refactor for optimize ibc channel, client related io (#1536)
Refactor in zoneconcierge to address inefficiencies and ambiguities in
the logic for retrieving IBC channels and clients

### 1. GetAllOpenZCChannels

- Currently, it retrieves all channels, but by using
`GetAllChannelsWithPortPrefix`, we can explicitly fetch only the
channels for the zoneconcierge port. This makes it more efficient and
ensures that channels from other ports are never fetched.

### 2. GetChannelForConsumer

- Currently written to iterate over all channels to obtain the channel
ID, but since RegisteredConsumer.metadata already contains the channel
ID explicitly, we can use that directly for a more efficient and
explicit retrieval.

### 3. GetAllConsumerIDs -> GetAllRegisteredCosmosConsumers
- At present, GetAllConsumerIDs is used to retrieve consumerIDs for
packet broadcasting, but to enable explicit channel ID access and more
efficient iteration, I’m adding `GetAllRegisteredCosmosConsumers`.
- But since the introduction of map in #1540, it is not currently used,
but kept for potential future use.